// frontend/src/interfaces/ClientData.tsx
export interface ClientData {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  birthdate: string;
  identificationNumber: string;
  address: string;
  workPlace: string;
  workPhone: string;
  startDate: string;
  workAddress: string;
  bossName: string;
  bossPhone: string;
  personalReference1: string;
  personalReferencePhone1: string;
  personalReference2: string;
  personalReferencePhone2: string;
  workReference1: string;
  workReferencePhone1: string;
  workReference2: string;
  workReferencePhone2: string;
}